ISPN-3674 LuceneStoreConfiguration should be renamed to LuceneLoaderConfiguration
