+++ lucene/lucene-v3/src/test/java/org/infinispan/lucene/cacheloader/LargeIndexesTest.java
@@ -1,4 +1,4 @@
-package org.infinispan.lucene.cachestore;
+package org.infinispan.lucene.cacheloader;
 
 import java.io.IOException;
